DialogHolder separated into own file
