+++ common/src/main/java/com/microsoft/identity/common/internal/ui/webview/challengehandlers/ClientCertAuthChallengeHandler.java
@@ -36,7 +36,6 @@ import android.security.KeyChainException;
 import android.webkit.ClientCertRequest;
 
 import androidx.annotation.NonNull;
-import androidx.annotation.Nullable;
 import androidx.annotation.RequiresApi;
 
 import com.microsoft.identity.common.R;
@@ -56,8 +55,6 @@ import com.yubico.yubikit.piv.PivSession;
 import com.yubico.yubikit.piv.Slot;
 import com.yubico.yubikit.piv.jca.PivPrivateKey;
 import com.yubico.yubikit.piv.jca.PivProvider;
-
-import net.jcip.annotations.ThreadSafe;
 
 import java.io.IOException;
 import java.security.Key;
@@ -135,127 +132,6 @@ public final class ClientCertAuthChallengeHandler implements IChallengeHandler<C
     }
 
     /**
-     * Builds and shows SmartcardDialog instances while keeping track of the current dialog being shown to the user.
-     */
-    @ThreadSafe
-    private class DialogHolder {
-        //The current dialog that is showing, if any.
-        private SmartcardDialog mCurrentDialog = null;
-
-        /**
-         * Build and show picker that prompts user to select a certificate for authentication.
-         * @param certList List of YubiKitCertDetails that contains cert details only pertinent to the cert picker.
-         * @param positiveButtonListener A PositiveButtonListener to be set for a SmartcardCertPickerDialog.
-         * @param negativeButtonListener A NegativeButtonListener to be set for a SmartcardCertPickerDialog.
-         * @param cancelCbaCallback      A Callback that holds code to be run when a YubiKey unexpectedly unplugs.
-         */
-        public synchronized void showCertPickerDialog(@NonNull final List<YubiKitCertDetails> certList,
-                                                      @NonNull final SmartcardCertPickerDialog.PositiveButtonListener positiveButtonListener,
-                                                      @NonNull final SmartcardCertPickerDialog.NegativeButtonListener negativeButtonListener,
-                                                      @NonNull final SmartcardCertPickerDialog.CancelCbaCallback cancelCbaCallback) {
-            final SmartcardCertPickerDialog certPickerDialog = new SmartcardCertPickerDialog(
-                    certList,
-                    positiveButtonListener,
-                    negativeButtonListener,
-                    cancelCbaCallback,
-                    mActivity);
-            //Show cert picker dialog.
-            showDialog(certPickerDialog);
-        }
-
-        /**
-         * Build and show PIN dialog that prompts user to type in their PIN for their YubiKey.
-         * @param positiveButtonListener A PositiveButtonListener to be set for a SmartcardPinDialog.
-         * @param negativeButtonListener A NegativeButtonListener to be set for a SmartcardPinDialog.
-         * @param cancelCbaCallback      A Callback that holds code to be run when a YubiKey unexpectedly unplugs.
-         */
-        public synchronized void showPinDialog(@NonNull final SmartcardPinDialog.PositiveButtonListener positiveButtonListener,
-                                               @NonNull final SmartcardPinDialog.NegativeButtonListener negativeButtonListener,
-                                               @NonNull final SmartcardPinDialog.CancelCbaCallback cancelCbaCallback) {
-            final SmartcardPinDialog pinDialog = new SmartcardPinDialog(
-                    positiveButtonListener,
-                    negativeButtonListener,
-                    cancelCbaCallback,
-                    mActivity);
-            //Show PinDialog, which should always be called after a positive button press.
-            mDialogHolder.showDialog(pinDialog);
-        }
-
-        /**
-         * Builds and shows dialog informing the user of an expected or unexpected error.
-         * @param titleStringResourceId String resource id of the title text.
-         * @param messageStringResourceId String resource id of the message text.
-         */
-        public synchronized void showErrorDialog(final int titleStringResourceId, final int messageStringResourceId) {
-            showDialog(new SmartcardErrorDialog(
-                            titleStringResourceId,
-                            messageStringResourceId,
-                            new SmartcardErrorDialog.PositiveButtonListener() {
-                                @Override
-                                public void onClick() {
-                                    //Reset currentDialog to null
-                                    mCurrentDialog = null;
-                                }
-                            },
-                            mActivity));
-        }
-
-        /**
-         * Dismisses current dialog, if one is showing.
-         */
-        public synchronized void dismissDialog() {
-            showDialog(null);
-        }
-
-        /**
-         * Shows provided SmartcardDialog if not null.
-         * Automatically dismisses existing dialog (if showing).
-         * @param dialog SmartcardDialog object to be shown.
-         */
-        public synchronized void showDialog(@Nullable final SmartcardDialog dialog) {
-            //Dismiss current dialog, if one is currently showing.
-            if (mCurrentDialog != null) {
-                mCurrentDialog.dismiss();
-            }
-            //Set current dialog, which could be null.
-            mCurrentDialog = dialog;
-            //If current dialog is not null, show.
-            if (mCurrentDialog != null) {
-                // dispatch to main thread if not on main thread
-                mCurrentDialog.show();
-            }
-        }
-
-        /**
-         * Informs if an existing dialog is currently showing.
-         * @return True if a SmartcardDialog is currently showing. False otherwise.
-         */
-        public synchronized boolean isDialogShowing() {
-            return (mCurrentDialog != null);
-        }
-
-        /**
-         * Runs the onCancelCbaCallback code for the current dialog.
-         * Used when YubiKey is unexpectedly disconnected from device.
-         */
-        public synchronized void onCancelCba() {
-            if (mCurrentDialog != null) {
-                mCurrentDialog.onCancelCba();
-            }
-        }
-
-        /**
-         * Sets error mode for an existing SmartcardPinDialog.
-         */
-        @RequiresApi(api = Build.VERSION_CODES.LOLLIPOP)
-        public synchronized void setPinDialogErrorMode() {
-            if (mCurrentDialog instanceof SmartcardPinDialog) {
-                ((SmartcardPinDialog) mCurrentDialog).setErrorMode();
-            }
-        }
-    }
-
-    /**
      * Callback which will contain code to be run upon creation of a PivSession instance.
      */
     public interface IPivSessionCallback {
@@ -273,7 +149,7 @@ public final class ClientCertAuthChallengeHandler implements IChallengeHandler<C
     public ClientCertAuthChallengeHandler(@NonNull final Activity activity) {
         mActivity = activity;
         //Create instance of DialogHolder.
-        mDialogHolder = new DialogHolder();
+        mDialogHolder = new DialogHolder(mActivity);
         //Create and start YubiKitManager for UsbDiscovery mode.
         //When in Usb Discovery mode, Yubikeys that plug into the device will be accessible
         // once the user provides permission via the Android permission dialog.
