+++ common/src/main/res/values/strings.xml
@@ -21,7 +21,7 @@
     <string name="smartcard_pin_dialog_error_message">The PIN you entered was incorrect</string>
     <string name="smartcard_pin_layout_edittext_hint">PIN</string>
 
-    <string name="smartcard_error_dialog_positive_button">Ok</string>
+    <string name="smartcard_error_dialog_positive_button">OK</string>
 
     <string name="smartcard_max_attempt_dialog_title">Try again later</string>
     <string name="smartcard_max_attempt_dialog_message">Too many failed attempts, try again later.</string>
